deinline qfb IO
